Show proper percentage when driving PWM motors

+++ src/kernel/drivers/Drv8833Driver.hpp
@@ -82,11 +82,11 @@ private:
 
         void drive(MotorPhase phase, double duty = 1) override {
             int dutyValue = in1Channel.maxValue() / 2 + (int) (in1Channel.maxValue() / 2 * duty);
-            Log.traceln("Driving motor %s on pins %d/%d at %.1f%%",
+            Log.traceln("Driving motor %s on pins %d/%d at %d%%",
                 phase == MotorPhase::FORWARD ? "forward" : "reverse",
                 in1Channel.pin,
                 in2Channel.pin,
-                duty * 100);
+                (int) (duty * 100));
 
             switch (phase) {
                 case MotorPhase::FORWARD:
